Uploading my first project
This project I called "Calculator" and its name is pretty representative. This is a GUI hand calculator Java application. You can do all four basic mathematical operations (addition, substraction, multipication and division) , it has a text field where the results are displayed and it basically works like you would expect it to (I designed it to be very similar to whichever kind of hand calculator aplication people are used to ; like for instance the standard windows calculator) . It is still work in progress but it is pretty functional on these four basic operations.