Uploading 3 Personal Projects
Queues Simulator is a Java console application that simulates a generic queue (people enter the queue and wait for the program/webpage/videogame or whatever to distribute the to servers or so on. I developed it for a college course and it gave me a better understanding of how threads work.

The "ToDo List" is my latest work, it is a JavaFX GUI application , that is intended to help the user organize his/her time by having digital memos (to-do's) all stacked in a place rather that on post-it notes or other non-digital methods. It has a ListView that displays the added to-do's , that will be saves on a file when the application is exited and loaded again in when it is reopened. It has a TextArea that displays a more detailed description of the task and a due date. I added controls and buttons to make it more functional ; it is not a pretty GUI app but it is functional. It is certainly still work in progress.

The Chess Game is pretty obvious a chess game implemented through a Java console application. I intended to make a GUI for it but my skills lacked a bit and decided to get more into JavaFX first and the come back to provide it with a graphical interface. It is currently a console application , but i would say it is pretty intuitive. It still has some bugs that I need to handle. Definetly work in prograss but is somehow functional.